HBSD: ftp/curl: Fixup .onion patch
We want people to still be able to reach HardenedBSD's infrastructure
over Tor, so we need to fixup the patch that removes the prohibition of
Tor Onion Services due to changes in libcurl's lib/hostip.c file.

This commit should serve as a reminder of how changes we humans make to
technology impact human rights efforts.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>