HBSD: Add "general compilation hardening" USES
This commit introduces a general catch all for compiler hardening flags
that aren't paired with a specific feature (unlike CFI, for example).
The USES for this is "hardcflags" and the USE_HARDENING flag is
"HARDCFLAGS".

We also enable this by default, though we will most likely need to fix a
number of ports (possibly a large number.) The biggest issue will likely
be the -Wformat-security compilation flag.

As of this writing, the HardenedBSD infrastructure lacks a server that
can perform experimental builds. So we need to apply a broad stroke
initially and resolve failed ports.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>